Move types into shared folder
