+++ libs/@blockprotocol/graph/src/internal/mutate-subgraph/edge.ts
@@ -1,12 +1,12 @@
 import { BaseUri } from "@blockprotocol/type-system/slim";
 
-import { EntityId } from "../../types/entity.js";
+import { EntityId } from "../../shared/types/entity.js";
 import {
   OntologyTypeRevisionId,
   OutwardEdge,
   Subgraph,
-} from "../../types/subgraph.js";
-import { Timestamp } from "../../types/temporal-versioning.js";
+} from "../../shared/types/subgraph.js";
+import { Timestamp } from "../../shared/types/temporal-versioning.js";
 import { isEqual } from "./is-equal.js";
 
 /**
