+++ libs/@blockprotocol/graph/src/shared/types/block-graph.ts
