+++ libs/@blockprotocol/graph/src/shared/types/entity.ts
