+++ libs/@blockprotocol/graph/src/shared/types/linked-aggregation.ts
