+++ libs/@blockprotocol/graph/src/shared/types/ontology/entity-type.ts
