+++ libs/@blockprotocol/graph/src/shared/types/ontology/property-type.ts
