+++ libs/@blockprotocol/graph/src/shared/types/subgraph.ts
