+++ libs/@blockprotocol/graph/src/shared/types/subgraph/edges/generic-outward-edge.ts
