+++ libs/@blockprotocol/graph/src/shared/types/subgraph/edges/kind.ts
