+++ libs/@blockprotocol/graph/src/shared/types/subgraph/edges/outward-edge.ts
