+++ libs/@blockprotocol/graph/src/shared/types/subgraph/edges/variants.ts
