+++ libs/@blockprotocol/graph/src/shared/types/subgraph/graph-resolve-depths.ts
