+++ libs/@blockprotocol/graph/src/shared/types/temporal-versioning.ts
@@ -2,7 +2,7 @@
  * Types used in embedding applications and blocks that support multi-axis temporal versioning schemes.
  */
 
-import { Subtype } from "../util.js";
+import { Subtype } from "../../util.js";
 
 /**
  * An ISO 8601 formatted timestamp string
