+++ libs/@blockprotocol/graph/src/stdlib/bound.ts
@@ -1,4 +1,7 @@
-import { TemporalBound, TimeInterval } from "../types/temporal-versioning.js";
+import {
+  TemporalBound,
+  TimeInterval,
+} from "../shared/types/temporal-versioning.js";
 
 export const compareBounds = (
   left: TemporalBound,
