+++ libs/@blockprotocol/graph/src/types.ts
@@ -1,7 +1,7 @@
-export * from "./types/block-graph.js";
-export * from "./types/entity.js";
-export * from "./types/file.js";
-export * from "./types/linked-aggregation.js";
-export * from "./types/ontology.js";
-export * from "./types/subgraph.js";
-export * from "./types/temporal-versioning.js";
+export * from "./shared/types/block-graph.js";
+export * from "./shared/types/entity.js";
+export * from "./shared/types/file.js";
+export * from "./shared/types/linked-aggregation.js";
+export * from "./shared/types/ontology.js";
+export * from "./shared/types/subgraph.js";
+export * from "./shared/types/temporal-versioning.js";
